Do not try to delete if we have no entry
If the entry we try to fetch is false anyway then it makes no sense to
actually do the delete. Saves some DB interaction.

+++ lib/private/Files/Cache/Cache.php
@@ -464,6 +464,11 @@ class Cache implements ICache {
 	 */
 	public function remove($file) {
 		$entry = $this->get($file);
+
+		if ($entry === false) {
+			return;
+		}
+
 		$sql = 'DELETE FROM `*PREFIX*filecache` WHERE `fileid` = ?';
 		$this->connection->executeQuery($sql, array($entry['fileid']));
 		if ($entry['mimetype'] === 'httpd/unix-directory') {
